pid_test

+++ main/imu.c
@@ -2,8 +2,8 @@
 
 extern struct AccelGyroData_int32_t OFFSET_RAW;
 extern uint8_t RC_CHECK;
-extern double G;
-extern double PI;
+extern float G;
+extern float PI;
 extern float accel_scale;
 extern float gyro_scale;
 extern int YAW_CAL;
